Revert "[Feat] Frame 관련 페이지 구조 분리 및 페이지 별 Metadata 설정"
