--- app/about/page.tsx
@@ -1,9 +1,0 @@
-import AboutPageContent from '@/app/about/AboutPageContent';
-
-export const metadata = {
-  title: '더 알아보기',
-};
-
-export default function AboutPage() {
-  return <AboutPageContent />;
-}
