+++ app/contact/page.tsx
@@ -1,8 +1,5 @@
 import ContactPageContent from '@/app/contact/ContactPageContent';
 
-export const metadata = {
-  title: '문의하기',
-};
 export default function ContactPage() {
   return <ContactPageContent />;
 }
