+++ app/layout.tsx
@@ -4,11 +4,8 @@ import { digitalix, pretendard } from '@/utils/font';
 import Header from '@/components/common/Header';
 
 export const metadata: Metadata = {
-  title: {
-    template: '%s | Time Film',
-    default: 'Time Film',
-  },
-  description: '우리의 추억을 담다.',
+  title: 'Create Next App',
+  description: 'Generated by create next app',
 };
 
 export default function RootLayout({
