+++ app/report/page.tsx
@@ -1,0 +1,5 @@
+import ReportPageContent from '@/app/report/ReportPageContent';
+
+export default function ReportPage() {
+  return <ReportPageContent />;
+}
