+++ app/result/page.tsx
@@ -2,57 +2,53 @@
 
 import PhotoFrame from '@/components/common/PhotoFrame';
 import { useFrameStore } from '@/stores/useFrameStore';
-import useSkinStore from '@/stores/useSkinStore';
-import { useProgressStore } from '@/stores/useProgressStore';
-import { useEffect, useRef } from 'react';
+import { COLORS } from '@/types/colors';
+import { useRef } from 'react';
 import SwiperType from 'swiper';
+import 'swiper/css';
 import Carousel from '@/components/common/Carousel';
 import { SwiperSlide } from 'swiper/react';
-import { COLORS } from '@/types/colors';
-import { SKINS } from '@/types/skins';
-import Image from 'next/image';
+import useStickerStore from '@/stores/useStickerStore';
+import { STICKERS } from '@/types/stickers';
 
-type BtnClickEventType = 'skin' | 'color';
-
-export default function EditPageContent() {
-  const swiperColorRef = useRef<SwiperType | null>(null);
-  const swiperSkinRef = useRef<SwiperType | null>(null);
+export default function Result() {
+  const swiperRef = useRef<SwiperType | null>(null);
+  const stickerSwiperRef = useRef<SwiperType | null>(null);
 
   const bgColor = useFrameStore(s => s.color);
   const setbgColor = useFrameStore(s => s.setColor);
+  const layout = useFrameStore(s => s.layout);
+  const stickers = STICKERS[layout];
 
-  const skin = useSkinStore(s => s.skin);
-  const setSkin = useSkinStore(s => s.setSkin);
+  const selectedSticker = useStickerStore(s => s.selectedSticker);
+  const setSelectedSticker = useStickerStore(s => s.setSelectedSticker);
 
-  const setStep = useProgressStore(state => state.setStep);
-
-  useEffect(() => {
-    setStep(3);
-  }, [setStep]);
-
-  const handleButtonClick = (type: BtnClickEventType, id: string, index: number) => {
-    if (type === 'color') {
-      setbgColor(id);
-      if (swiperColorRef.current) {
-        swiperColorRef.current.slideToLoop(index, 300);
-      }
-    } else if (type === 'skin') {
-      setSkin(id);
-      if (swiperSkinRef.current) {
-        swiperSkinRef.current.slideToLoop(index, 300);
-      }
+  const handleColorButtonClick = (colorId: string, index: number) => {
+    setbgColor(colorId);
+    if (swiperRef.current) {
+      swiperRef.current.slideToLoop(index, 300);
     }
   };
 
-  const colorInitialSlide = COLORS.findIndex(c => c.id === bgColor);
-  const skinInitialSlideSkin = SKINS.findIndex(s => s.id === skin);
+  const handleStickerButtonClick = (stickerId: string, index: number) => {
+    setSelectedSticker(stickerId);
+    if (stickerSwiperRef.current) {
+      stickerSwiperRef.current.slideToLoop(index, 300);
+    }
+  };
+
+  const initialColorSlide = COLORS.findIndex(c => c.id === bgColor);
+  const initialStickerSlide = COLORS.findIndex(c => c.id === selectedSticker);
 
   return (
     <div className="flex flex-col w-full items-center">
+      <div className="absolute top-4 inset-x-4 bg-amber-950 h-20 flex justify-center items-center text-white font-bold">
+        Step Indicator
+      </div>
       <div className="flex flex-col gap-5 w-full items-center">
         <div className="flex flex-col gap-2 w-[70%] text-center pt-10">
           <p>프레임 색상</p>
-          <Carousel swiperRef={swiperColorRef} initialSlide={colorInitialSlide}>
+          <Carousel swiperRef={swiperRef} initialSlide={initialColorSlide}>
             {COLORS.map((c, index) => (
               <SwiperSlide key={c.id}>
                 <div className="flex justify-center items-center">
@@ -61,7 +57,7 @@ export default function EditPageContent() {
                   >
                     <button
                       className={`w-8 h-8 rounded-full ${c.color}`}
-                      onClick={() => handleButtonClick('color', c.id, index)}
+                      onClick={() => handleColorButtonClick(c.id, index)}
                     />
                   </div>
                 </div>
@@ -70,25 +66,19 @@ export default function EditPageContent() {
           </Carousel>
         </div>
         <div className="flex flex-col gap-2 w-[70%] text-center">
-          <p>프레임 스킨</p>
-          <Carousel swiperRef={swiperSkinRef} initialSlide={skinInitialSlideSkin}>
-            {SKINS.map((s, index) => (
+          <p>스티커</p>
+          <Carousel swiperRef={stickerSwiperRef} initialSlide={initialStickerSlide}>
+            {stickers.map((s, index) => (
               <SwiperSlide key={s.id}>
                 <div className="flex justify-center items-center">
                   <div
-                    className={`w-12 h-12 bg-white/50 rounded-full border-2 flex justify-center items-center ${skin === s.id ? 'border-black' : 'border-transparent'}`}
+                    className={`w-12 h-12 bg-white/50 rounded-full border-2 flex justify-center items-center ${selectedSticker === s.id ? 'border-black' : 'border-transparent'}`}
                   >
                     <button
                       className={`w-8 h-8 rounded-full`}
-                      onClick={() => handleButtonClick('skin', s.id, index)}
+                      onClick={() => handleStickerButtonClick(s.id, index)}
                     >
-                      <Image
-                        src={s.icon || '/images/icon/default-icon.png'}
-                        alt={s.id}
-                        width={32}
-                        height={32}
-                        className="rounded-full"
-                      />
+                      {s.id}
                     </button>
                   </div>
                 </div>
